Moved Terminus::confirm into Input helper class
