+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 ### Changed
 - Extricated the request logic from TerminusCommand class and moved it to the Request class. (#704)
 - Replaced Mustache templates with Twig. (#730)
+- Resolved Terminus base class' and Runner's interfunctionality. Terminus can now be used as a library. (#738)
 
 ##[0.9.3] - 2015-11-17
 ### Added
